requirements.txt
fastapi
uvicorn
pandas
redis
openai
python-dotenv
openpyxl